Integrate link tool into same pattern as others
